+++ addons/api/addon/abilities/scope.js
@@ -25,4 +25,42 @@ export default class ScopeAbility extends ModelAbility {
   get canDetachStoragePolicy() {
     return this.hasAuthorizedAction('detach-storage-policy');
   }
+
+  /**
+   * Setting the alias target suffix is allowed on project scopes when the
+   * `set-alias-target-suffix` authorized action is present.
+   * @type {boolean}
+   */
+  get canSetAliasSuffix() {
+    return (
+      this.model?.isProject &&
+      this.hasAuthorizedAction('set-alias-target-suffix')
+    );
+  }
+
+  /**
+   * Reading the alias target suffix is allowed on project scopes
+   * when the `get-alias-target-suffix` authorized action is present.
+   * @type {boolean}
+   */
+  get canGetAliasSuffix() {
+    return (
+      this.model?.isProject &&
+      this.hasAuthorizedAction('get-alias-target-suffix')
+    );
+  }
+
+  /**
+   * Removing the alias target suffix is allowed on project scopes when the
+   * `remove-alias-target-suffix` authorized action is present and a suffix
+   * is currently set.
+   * @type {boolean}
+   */
+  get canRemoveAliasSuffix() {
+    return (
+      this.model?.isProject &&
+      Boolean(this.model?.alias_suffix) &&
+      this.hasAuthorizedAction('remove-alias-target-suffix')
+    );
+  }
 }
